backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
pandas==2.2.2
mongoengine==0.29.1
requests==2.31.0
python-dotenv==1.0.1
yfinance
google-genai
